Users and user errors tables and their accessors declaration

+++ appdatabase.cpp
@@ -10,11 +10,15 @@ AppDatabase::AppDatabase(std::string name)
     countTickets();
 }
 
-int AppDatabase::getTicketsAmount()
+int AppDatabase::getTicketsABamount()
 {
-    return ticketsAmount;
+    return ticketsABamount;
 }
 
+int AppDatabase::getTicketsCDamount()
+{
+    return ticketsCDamount;
+}
 
 AppDatabase::~AppDatabase()
 {
@@ -42,15 +46,37 @@ bool AppDatabase::initSQLite(std::string name)
 bool AppDatabase::createTables()
 {
     try {
-        db->exec("CREATE TABLE IF NOT EXISTS questions ("
-                 "id INTEGER PRIMARY KEY,"
-                 "ticket_num INTEGER,"
-                 "question_num INTEGER,"
-                 "image TEXT,"
-                 "question TEXT,"
-                 "answers TEXT,"
-                 "comment TEXT,"
-                 "right_answer INTEGER);");
+        db->exec(
+            "CREATE TABLE IF NOT EXISTS questions ("
+            "id INTEGER PRIMARY KEY,"
+            "category INTEGER,"
+            "ticket_num INTEGER,"
+            "question_num INTEGER,"
+            "image TEXT,"
+            "question TEXT,"
+            "answers TEXT,"
+            "comment TEXT,"
+            "right_answer INTEGER);"
+        );
+
+        db->exec(
+            "CREATE TABLE IF NOT EXISTS users ("
+            "login TEXT PRIMARY KEY,"
+            "password TEXT,"
+            "name TEXT,"
+            "permissions INTEGER);"
+        );
+
+        db->exec(
+            "CREATE TABLE IF NOT EXISTS users_errors ("
+            "question_id INTEGER,"
+            "user_id INTEGER,"
+            "answer INTEGER,"
+            "FOREIGN KEY(question_id) REFERENCES questions(id),"
+            "FOREIGN KEY(user_id) REFERENCES users(id));"
+        );
+
+
         return true;
     } catch(std::exception &e) {
         std::cerr << "exeption: " << e.what() << std::endl;
@@ -61,10 +87,17 @@ bool AppDatabase::createTables()
 void AppDatabase::countTickets()
 {
     try {
-        SQLite::Statement  query(*db, "SELECT MAX(ticket_num) FROM questions");
+        SQLite::Statement queryAB(*db, "SELECT MAX(ticket_num) FROM questions WHERE category = " + std::to_string(Question::Category::AB));
 
-        query.executeStep();
-        ticketsAmount = query.getColumn(0).getInt();
+        queryAB.executeStep();
+        ticketsABamount = queryAB.getColumn(0).getInt();
+        queryAB.reset();
+
+        SQLite::Statement queryCD(*db, "SELECT MAX(ticket_num) FROM questions WHERE category = " + std::to_string(Question::Category::CD));
+
+        queryCD.executeStep();
+        ticketsABamount = queryCD.getColumn(0).getInt();
+        queryCD.reset();
 
     } catch(std::exception &e) {
         std::cerr << "exeption: " << e.what() << std::endl;
@@ -75,22 +108,26 @@ void AppDatabase::countTickets()
 bool AppDatabase::addQuestion(Question &question)
 {
     try {
-        if (question.question_num >= 100 or question.question_num < 1 or question.ticket_num < 1)
-            return false;
+        if (question.question_num >= 100 or
+            question.question_num < 1 or
+            question.ticket_num < 1 or
+            question.category < 0
+        ) return false;
 
-        uint id = question.ticket_num + question.question_num*100;
+        uint id = (question.ticket_num + question.question_num*100)*10 + question.category;
 
         // Prepare query
-        SQLite::Statement query {*db, "INSERT OR IGNORE INTO questions VALUES (?,?,?,?,?,?,?,?)"};
+        SQLite::Statement query {*db, "INSERT OR IGNORE INTO questions VALUES (?,?,?,?,?,?,?,?,?)"};
 
         query.bind(1, id);
-        query.bind(2, question.ticket_num);
-        query.bind(3, question.question_num);
-        query.bind(4, question.image_base64);
-        query.bind(5, question.question_text);
-        query.bind(6, question.answers);
-        query.bind(7, question.comment);
-        query.bind(8, question.rightAnswer);
+        query.bind(2, question.category);
+        query.bind(3, question.ticket_num);
+        query.bind(4, question.question_num);
+        query.bind(5, question.image_base64);
+        query.bind(6, question.question_text);
+        query.bind(7, question.answers);
+        query.bind(8, question.comment);
+        query.bind(9, question.rightAnswer);
 
         query.exec();
 
@@ -107,13 +144,17 @@ bool AppDatabase::addQuestion(Question &question)
 bool AppDatabase::modifyQuestion(Question &question)
 {
     try {
-        if (question.question_num >= 100 or question.question_num < 1 or question.ticket_num < 1)
-            return false;
+        if (question.question_num >= 100 or
+            question.question_num < 1 or
+            question.ticket_num < 1 or
+            question.category < 0
+        ) return false;
 
-        uint id = question.ticket_num + question.question_num*100;
+        uint id = (question.ticket_num + question.question_num*100)*10 + question.category;
 
         // Prepare query
         SQLite::Statement query {*db, "UPDATE questions SET "
+                                      "category = :category,"
                                       "ticket_num = :ticket_num,"
                                       "question_num = :question_num,"
                                       "image = :image,"
@@ -123,6 +164,7 @@ bool AppDatabase::modifyQuestion(Question &question)
                                       "right_answer = :right_answer "
                                       "WHERE id = " + std::to_string(id)};
 
+        query.bind(":category", question.category);
         query.bind(":ticket_num", question.ticket_num);
         query.bind(":question_num", question.question_num);
         query.bind(":image", question.image_base64);
@@ -140,13 +182,16 @@ bool AppDatabase::modifyQuestion(Question &question)
     }
 }
 
-bool AppDatabase::deleteQuestion(int ticketNum, int questionNum)
+bool AppDatabase::deleteQuestion(Question::Category category, int ticketNum, int questionNum)
 {
     try {
-        if (questionNum >= 100 or questionNum < 1 or ticketNum < 1)
-            return false;
+        if (questionNum >= 100 or
+            questionNum < 1 or
+            ticketNum < 1 or
+            category < 0
+        ) return false;
 
-        uint id = ticketNum + questionNum*100;
+        uint id = (ticketNum + questionNum*100)*10 + category;
 
         SQLite::Statement query {*db, std::string("DELETE FROM questions WHERE id = " + std::to_string(id))};
 
@@ -159,16 +204,19 @@ bool AppDatabase::deleteQuestion(int ticketNum, int questionNum)
     }
 }
 
-std::shared_ptr<Question> AppDatabase::getQuestion(int ticketNum, int questionNum)
+std::shared_ptr<Question> AppDatabase::getQuestion(Question::Category category, int ticketNum, int questionNum)
 {
     try {
-        if (questionNum >= 100 or questionNum < 1 or ticketNum < 1)
-            return nullptr;
+        if (questionNum >= 100 or
+            questionNum < 1 or
+            ticketNum < 1 or
+            category < 0
+        ) return nullptr;
 
-        std::shared_ptr<Question> newQuestion {new Question};
+        std::shared_ptr<Question> requestedQuestion {new Question};
 
-        uint id = ticketNum + questionNum*100;
-        newQuestion->question_num = id;
+        uint id = (ticketNum + questionNum*100)*10 + category;
+            requestedQuestion->question_num = id;
 
         SQLite::Statement query {*db, std::string("SELECT * FROM questions WHERE id = " + std::to_string(id))};
 
@@ -177,16 +225,16 @@ std::shared_ptr<Question> AppDatabase::getQuestion(int ticketNum, int questionNu
         }
         while (query.executeStep())
         {
-            newQuestion->ticket_num = query.getColumn(1).getInt();
-            newQuestion->question_num = query.getColumn(2).getInt();
-            newQuestion->image_base64 = query.getColumn(3).getString();
-            newQuestion->question_text = query.getColumn(4).getString();
-            newQuestion->answers = query.getColumn(5).getString();
-            newQuestion->comment = query.getColumn(6).getString();
-            newQuestion->rightAnswer = query.getColumn(7).getInt();
+                requestedQuestion->ticket_num = query.getColumn(1).getInt();
+                requestedQuestion->question_num = query.getColumn(2).getInt();
+                requestedQuestion->image_base64 = query.getColumn(3).getString();
+                requestedQuestion->question_text = query.getColumn(4).getString();
+                requestedQuestion->answers = query.getColumn(5).getString();
+                requestedQuestion->comment = query.getColumn(6).getString();
+                requestedQuestion->rightAnswer = query.getColumn(7).getInt();
         }
 
-        return newQuestion;
+        return  requestedQuestion;
     } catch(std::exception &e) {
         std::cerr << "exeption: " << e.what() << std::endl;
         return nullptr;
